Add additional file error types
When a file being read or written to. There are no exceptions in
the case where a file is not writable or readable.

Not being writable or readable should be considered a checked error

Add not being writable and readable as custom duke exceptions.

Making custom exceptions allow for custom behavior for these errors
that align with the requirements of the project