Add file handler class
Reading and writing to files have to be done separately in each class
that uses it.

The reading and writing to files can be abstracted to avoid code
duplication.

Create a file handler class to abstract these details. Modify he
exceptions handling the read and write permissions to be more helpful.

Less boilerplate needs to be written for file reading and writing.