Add the ability to save and load
Users had to enter all the task details and this was not persistent
across user sessions.

Add the ability to save and load. Save files have the format of tasks
being separated by newlines and different attributes of tasks are
separated by commas. The program will parse and unparse this when
saving and loading. Saving and loading is done via the SaveHandler
class. Classes that can be saved implements the Saveable interface. Add
the SaveData class to contain the data that needs to be saved.
Update the text-ui-test tests. Update .gitignore to account for the
new save files. Add InavlidSaveException class to handle exceptions for
invalid formatting of the save file.

The SaveHandler class is to abstract away saving. The SaveData class
is to encapsulate the test data to make updating it in the foture
easier.