Abstract parsing logic into Parser class and make Command classes
Parsing/command logic is in the class of the chatbot.

This makes the responsibility of both parsing and command logic
fall into the chatbot's hands. This also makes parsing and logic highly
coupled with the chatbot.

Let's do,
- Abstract command logic into their corresponding class
- Make Command as a class for shared logic amongst commands
- Abstract parsing logic into its own class
- Fix other minor issues
- Remove unnecessary methods from chatbot class
- Account for command and parsing logic in the chatbot class

The parser should return a Command object and the chatbot should
execute the command. This completely separates their logic.
If there is an error with the parsing, it has to do with the format
of the argument. Whereas if there is a problem with executing the
command, is has not much to do with how the user formatted their
argument. This separates the logic and reduces coupling as well.