Divide classes into packages
Classes are in the source folder without being associated with a
package other than the classes in the exceptions folder.

This is poor organisation and quite messy.

Divide classes into folders that make sense:
- Various tasks classes into the tasks folder
- Ui components go into the ui folder
- Utility functions and classes go into utils folder
- Save and storage classes go into the storage folder
- Exceptions go into the exceptions folder
- The various commmand classes go into the Command folder
Also, change cases to accommodate new file paths.

Improves maintainability, readability, scalability as well
as code reusability. This is a relatively smaller project so do not
need that many nested folders.