Change name of stringIsInt method to stringIsIndex
Method that checks if a string is an index of an array is called
stringIsInt.

This is inaccurate as it an integer can be less than 0 which is not
under the scope of the method.

Change name to stringIsIndex instead to more accurately reflect its
function.

This should help me and developers in the future understand and use
the function.