Add find command
Users cannot search for tasks.

Users should be able to search for tasks using a command.

Do:
- FindCommand class for the find command
- Change error logic for TaskList class
- Change listTasks method to work for both find and list commands and
update Junit tests
- Add to text ui tests
- Add find command to parser
- Add nameContainsString method to Task class

Use taskList to store the search results as well as the purpose of
the TaskList class in the first place is to store tasks and manipulate
them. The same principle can be applied to the find command as well.