Fix bugs caused by last merge commit and fix coding standards.
Errors due to merge commit exist. Coding standards are not up to par
in other branches that were merged.

Important to fix the bugs and ensure coding standards are acceptable.

Fix residual merge conflict in UtilsTest class. Fix coding standards
such as wrong import order and indentation of case statements. Add
Javadocs to methods that did not have Javadocs.

Fix it in a separate/branch and commit because the last commit was
about merging.