Add notes as a feature
User has no easy way to input notes for each task.

It would be convenient for the user to input notes to remember
specific things about each task.

Add:
- The ability to save notes to a particular task.
- The ability to get the notes to a particular notes.
- The ability to load notes from the save file.
- The ability to save notes to the save file.

Add new commands and lines to the parser and format that follows
the same format as the rest of the commands to keep it consistent.