Add and throw CommandExceptions in various methods
The exceptions of the user entering the wrong format of the command
is not differentiated from the exceptions when the user enters invalid
values.

It makes sense to abstract them as different errors and signal them
as different errors for both for the users and the developers.

Make new kind of Exception called CommandException. This exception
is thrown when a Command execution fails for whatever reason.

This is to differentiate formatting errors of the user from the
wrong command inputs of the user and various other errors. Again, I
do believe this should help the user and developer.